Define public API
